pyproject.toml
[project]
name = "gza-agent"
dynamic = ["version"]
description = "A coding AI agent runner for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Matthew Hawthorne", email = "mh@supremeinformatics.com" }
]
keywords = ["claude", "ai", "agent", "runner", "cli", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/mhawthorne/gza"
Repository = "https://github.com/mhawthorne/gza"
Issues = "https://github.com/mhawthorne/gza/issues"

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0",
    "pytest>=8.0",
    "build>=1.0",
    "twine>=5.0",
]

[project.scripts]
gza = "gza.cli:main"

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gza"]

[tool.hatch.version]
source = "vcs"
fallback-version = "0.1.0.dev0"

[tool.hatch.build.hooks.vcs]
version-file = "src/gza/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: mark test as integration test (requires credentials and Docker)",
]
